Enable check for implicit casts
- Fix syntax for upgrading errors in the analysis options
- Ignore todo hints since they are very noisy
- Enable check for implicit casts
- Fix a bunch of implicit casts in the test directory

+++ test/runner/hybrid_test.dart
@@ -268,7 +268,7 @@ void main() {
 
       // Expect that the socket disconnects at some point (presumably when the
       // isolate closes).
-      var port = await channel.stream.first;
+      var port = await channel.stream.first as int;
       var socket = await Socket.connect("localhost", port);
       expect(socket.listen(null).asFuture(), completes);
 
@@ -294,7 +294,7 @@ void main() {
 
       // Expect that the socket disconnects at some point (presumably when the
       // isolate closes).
-      var port = await channel.stream.first;
+      var port = await channel.stream.first as int;
       var socket = await Socket.connect("localhost", port);
       expect(socket.listen(null).asFuture(), completes);
       channel.sink.add(null);
